Implement materials, Vec3 operation, use Box instead of refs
Added Material fields to HitRecord and Sphere

Use Lambertian with two new objects

Implement Metal material

Add reflect() to reflect a vector along a normal

Implement the element wise multiplication between Vec3

Switch from references to Box and Rc:

References were creating a great deal of overhead on first
instantiation and lifetimes in the form of <'a> was getting complex.
Using Boxes to allocate on the heap and Rc (reference count) on
various fields simplifies a great deal.